requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.32.3

# AI/ML Stack
groq>=0.11.0
sentence-transformers==3.1.1
jina-embeddings>=0.2.0
torch==2.4.1
torchvision==0.19.1
transformers==4.40.0
scikit-learn==1.4.2
numpy<2

# Vector Database
qdrant-client==1.17.1

# Document Processing
pypdf2==3.0.1
pytesseract==0.3.13
pillow==10.4.0